Add in git/specs.md based on a conversation from Slack.
Normally, this would be in a GitHub issue, but we don't want this information to be public.
Part of the protection here is that the student's don't realize the process we are doing
to verify their commits which removes their ability to creatively think around the solution.
We want them creatively figuring out how to build a chess client, not getting away without
writing git commits.