Move Phase configuration data into their own records...
This shifts the emphasis on data from specific fields to the single object 'Phase'.
This makes it easier to add new phases since all required fields are required to
construct the appropriate 'Phase' object.

This sets the stage for two other adjustments:
1. Obviating and removing the PhaseUtils class
2. Loading in the list of phases dynamically
   instead of changing code files to adjust
   the constants.

This isn't the final step because we can't dynamically
generate 'enum' types, but this is the first step
in a series of transformations that will end up
simplifying the responsibilites of course maintainers
in the future.