Only reveal commit verification results on a successful submission
In discussion with @19mdavenport, we decided that this will allow
us to keep our cards in our hand until they need to be revealed.

Since we only pay attention to the commit history for passing submissions,
it makes sense to only show this feedback on a passing submission as well.