Increase line change limit to 20
This choice was made intentionally with research and discussion
with @19mdavenport. Several students were randomly sampled from
the previous semester, and their git histories were analyzed.

Of the 7 students analyzed, they could have all passed if this limit
had been in force. 20 changes could be 10 insertions and 10 deletions.
That is basically one method. The idea is that each method can be written
and committed, and that will result in a passing history.

Of the commits analyzed, ~27% would be skipped because of this threshold,
but also each student had ~50% more commits than the minimum requirement.

Teaching students to author commits intentionally and frequently, with
all related changes together will lead to passing the history, and if
edge cases ever arise, the submission can always be manually approved.