Apply git penalties before checking if it lowers the score
Previously, it would be possible for a later submission to lower a student's score if it were close to the threshold,
and then the commit verification penalty were applied on top. This is a more accurate calculation.

It also aligns with the student's needs. This code will only matter after a student's submission has been rejected,
then manually approved, and then submitted again for additional points. If the new points minus the penalty doesn't help,
then they wouldn't want to lower their score.