Remove Canvas requirement that RubricItems have positive points
The Git commit verification system intentionally places
negative valued rubric items in the rubrics as penalties
when the conditions are not met properly.

It is unclear why this condition was originally inserted,
but in discussion with @19mdavenport, we decided it would
safe to remove this. In fact, it was commented that it may
actually be better to remove this check than to leave
it in.