Improve documentation and code structure in Scorer...
- Add or adjust JavaDoc comments on several key methods
- Extract code into transformRubric() method
- Convert code to returning immediately when the Submission is generated.
  Though the single-exit point code style is sometimes better,
  In this case there are no resoruces to consider, and I
  think it's clearer to emphasize that when the submission is generated,
  this function's processing is DONE.

Also note that during this process, I confirmed the call signatures of all methods
in this class. Notably, I considered splitting code from Scorer.score()
into a separate method that would just return a String field representing the notes.
While this woul remove the nearly duplicate calls to generateSubmissionObject(),
it doesn't account for the adjustments made to the Rubric when the missing values
are populated from Canvas. I decided it wasn't worth adjusting now, although the
architecture of the methods may appreciate revisiting again when new functionality
is requested.