Process each unique submission only once
When the same commit is resubmitted for the same phase,
this generates a new submission, but is not meaningful
for our purposes. Since the SubmissionDAO#manuallyApproveSubmission()
method already modifies all submissions with the same identifying
information, this resulted in us writing to the database many
extra times.

Ideally, we would have a Submission ID to track them by individually,
but we do not. However, the Submission#equals() method is defined
the same as the way this behavior works, so we can rely on set
equality to give us the answer we expect.