Add special phase for GitCommits which can be represented by the GradingRubric.
Unfortunately, this requires adding the option in several places to
satisfy the completness of all switch cases.

We considered putting the config under the 'Quality' phase, and the
GitCommits RubricType. This actually seems like a great approach,
but it misuses the phase name. There are tradeoffs all the way around.

The phase is named 'Commits' because SQL requires the name to be less
than 9 chars which rules out 'GitCommits'.