Streamline handling of all items within Rubrics
This creates levereages default implementations in RubricConfigDao
to avoid duplicating the same logic in all DAO implementations.

It also greates two new getters allConfigItems() and allRubricItems()
on the appropriate classes. This prepares preparation improves code
readability of classes that perform operations on all of the items,
and also protects against change when items are added/removed to the
record.