CommitAnalytics will continue grading without killing the observer on error...
Specifically, if a student has nuked their repo, then they clone it and restart with
a new commit history, throwing away the old commits. This isn't necessary, and isn't
the best approach, but it's what they do. The AutoGrader would then fail to find
the previous submission head hash. This error would end up calling
Observer.notifyError() which is apparently terminal to to autograding.

By adjusting the handling of the error, we correctly allow the grading
to continue and merely provide a 'failed' status that a TA can manually approve.