Correct issue in reasonable performanc test...
Previously, if run within the first 200 minutes of the day (~3 hrs),
the test would fail because it would backdate the commits onto a separate
date. This changes the behavior to always expect two days and thus avoid the issue.

Note that other tests generating lots of commits may face similar
issues when executing the tests near midnight.