Move TrafficController into controller package
This is the only current 'controller' file outside
of the 'controller' package. This statement was verified
by searching for the following regex in the project:
package [\w.]*controller

Additionally, this move strengthens the overall
layered architecture of the autograding system.
The 'controller' package depends on the 'autograder'
package, but there are no dependencies in the reverse
direction.