Clean up error handling of SubmissionUtils:getRemoteHeadHash()
It was still throwing RuntimeExceptions
from the days before we realized the benefits
of typed exceptions.

It was also double logging the error,
which is avoided now by merely throwing the error
and allowing the more central location to
LOGGER.error() the message.

In making this change, we recurisvely
verified all other calls from the method
to ensure that no other errors under this
call tree are being thrown as RuntimeExceptions.
There were no other instances here.