Simplfy usages: Avoid async...
The 'async' keyword is super helpful when dealing with
concurrent programming and requests that take awhile.
However, now that all these usages are merely simple
wrappers around the actual methods, it is no longer
helpful to use async/await. In fact, requiring that
syntax only makes it easier for a future developer
to accidentally leave off an important 'await' and
cause requests to be strangely completing in the
background.

Principle: If something can be simplified: do it!