Move controller/*Exception.java classes into a sub-package...
This is a novel move in this repository, and it is not
applied to other packages as well. In this case, there
are a large number of controller classes intermingled
with multiple exception classes which warrants additional
organizational layers. Other packages that have *both*
fewer classes and fewer exceptions are fine to remain.