Move cleanup responsibility into GitHelper:fetchRepoFromUrl(String)...
This simplifies the usage of this function and makes it more friendly
to reuse in other contexts. It also makes more sense since, with this
simple overload, we just want it to work. The actual implementation
has not changed and has the same behavior as before, allowing more
advanced usage if desried.