Invert dependencies for Scorer and LateDayCalculator...
The objects are now passed between classes as constructor
dependencies, which will allow for data to be shared and
preserved among queries requesting the same information.

This addresses #508 but does not completely resolve the issue.

As part of this change, the instantiation of default dependencies
was removed in favor of intentially pushing the instatiation of
the objects up into a central location.