Fix flakey commit ammend test...
'Flakey' is a strong word because the test did nearly all
of the time it ran.

Detecting an ammended commit requires an exact timestamp match.
Git timestamps have second precision.
Without actually using the '--amend' flag,
the test would create multiple commits in quick succession
which is effectively the same as ammending commits.
As long as all the commits are authored in the same second,
the test works.

However, there is a non-zero chance that the span of commits
would cross the boundary of a second and cause the test to fail.
The exact probability has to do with the execution time of
the machine creating the commits, but the key point is that
the probability is non-zero.

This change leverages a different overload of GitHelperUtils.makeCommit()
to specify an exact Instant where the commit should be authored.
We compute 'now()' once, and then determine the timestamps
of the other commits based on subtractions from now().
This guarantees us that the commits will *always* have
the timestamps we intend for them to have.