Avoid test name key collisions...
It is possible for students to *errorenously* create multiple tests
with the same name. The tests would evaluate correctly, but then
the TestAnalyzer would only given credit for each *unique*
test name. When students leave the same name for multiple tests,
even if the function name is different, then this results in
undercounting the student's tests.

This commit guarantees that the keys used to insert into the Map
are always unique by adding an incrementing number to the end.
We are careful to only affect the key as it goes into the Map,
and not the testName which will be displayed to the user.

An alternative approach, and maybe more complete, would be to
refactor the Map<String, TestNode> into a List<TestNode>.
This would more faithfully represent each test just simply
for its existence. There is no need to look up each
TestNode by the ID. However, this would be an extensive
refactor since the TestNode class is used in many places
in the AutoGrader and even gets serialized and sent to the
front end.

Therefore, this change of guaranteeing that the Map test keys
are unique is the minimally invasive approach to solve the problem.