Rename testdb.env to testdb.example.env
We could potentially have this be testdb.env, but I want to ensure that any future autograder devs don't commit their MySQL keys into the repository. There are commands to ensure that changes to the testdb.env aren't staged, but I don't think it's fair to expect future TAs to know them, and this solution is easier then documenting the git commands.